requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
